+++ web/usuaris/src/__tests__/molecules/DeleteConfirmationModal.test.tsx
@@ -3,19 +3,25 @@ import { server } from '../../__mocks__/server'
 import { DeleteConfirmationModal } from '../../components/molecules/DeleteConfirmationModal'
 import { fireEvent, render, screen, waitFor } from '../test-utils'
 
-const defaultProps = {
+const defaultDeleteUser = {
   open: true,
   toggleModal: vi.fn(),
   idToDelete: '1',
 }
 
+const defaultDeleteMultipleUsers = {
+  open: true,
+  toggleModal: vi.fn(),
+  idsToDelete: ['1', '3', '4'],
+}
+
 describe('DeleteConfirmationModal', () => {
   it('renders correctly', async () => {
-    render(<DeleteConfirmationModal {...defaultProps} />)
+    render(<DeleteConfirmationModal {...defaultDeleteUser} />)
 
     await waitFor(() => {
       expect(
-        screen.getByText('Estàs segur que vols eliminar aquest usuari?')
+        screen.getByText('Estàs segur que vols eliminar aquest/s usuari/s?')
       ).toBeInTheDocument()
       expect(screen.getByTestId('confirm-button')).toBeInTheDocument()
       expect(screen.getByTestId('cancel-button')).toBeInTheDocument()
@@ -25,12 +31,12 @@ describe('DeleteConfirmationModal', () => {
     fireEvent.click(cancelButton)
 
     await waitFor(() => {
-      expect(defaultProps.toggleModal).toHaveBeenCalled()
+      expect(defaultDeleteUser.toggleModal).toHaveBeenCalled()
     })
   })
 
   it('handles successful deletion', async () => {
-    render(<DeleteConfirmationModal {...defaultProps} />)
+    render(<DeleteConfirmationModal {...defaultDeleteUser} />)
 
     await waitFor(() => {
       const confirmButton = screen.getByTestId('confirm-button')
@@ -39,15 +45,15 @@ describe('DeleteConfirmationModal', () => {
 
     await waitFor(() => {
       expect(
-        screen.getByText('Usuario eliminado correctamente')
+        screen.getByText('Usuari/s eliminat/s correctament')
       ).toBeInTheDocument()
-      expect(defaultProps.toggleModal).toHaveBeenCalled()
+      expect(defaultDeleteUser.toggleModal).toHaveBeenCalled()
     })
   })
 
   it('handles deletion error', async () => {
     server.use(...errorHandlers)
-    render(<DeleteConfirmationModal {...defaultProps} />)
+    render(<DeleteConfirmationModal {...defaultDeleteUser} />)
 
     await waitFor(() => {
       const confirmButton = screen.getByTestId('confirm-button')
@@ -56,7 +62,43 @@ describe('DeleteConfirmationModal', () => {
 
     await waitFor(() => {
       expect(
-        screen.getByText(/Error en eliminar l'usuari/i)
+        screen.getByText(`Error en eliminar l'usuari/s`)
+      ).toBeInTheDocument()
+    })
+  })
+
+  it('handles multiple deletion successfully', async () => {
+    render(<DeleteConfirmationModal {...defaultDeleteMultipleUsers} />)
+
+    await waitFor(() => {
+      const confirmButton = screen.getByTestId('confirm-button')
+      fireEvent.click(confirmButton)
+    })
+
+    await waitFor(
+      () => {
+        expect(
+          screen.getByText('Usuari/s eliminat/s correctament')
+        ).toBeInTheDocument()
+
+        expect(defaultDeleteMultipleUsers.toggleModal).toHaveBeenCalled()
+      },
+      { timeout: 5000 }
+    )
+  })
+
+  it('handles multiple deletion error', async () => {
+    server.use(...errorHandlers)
+    render(<DeleteConfirmationModal {...defaultDeleteMultipleUsers} />)
+
+    await waitFor(() => {
+      const confirmButton = screen.getByTestId('confirm-button')
+      fireEvent.click(confirmButton)
+    })
+
+    await waitFor(() => {
+      expect(
+        screen.getByText(`Error en eliminar l'usuari/s`)
       ).toBeInTheDocument()
     })
   })