+++ web/usuaris/src/hooks/index.ts
@@ -1,5 +1,7 @@
+export { useDeleteMultipleUsers } from './useDeleteMultipleUsers'
 export { useDeleteUser } from './useDeleteUser'
 export { useGetItineraries } from './useGetItineraries'
 export { useGetUsers } from './useGetUsers'
 export { useLogin } from './useLogin'
 export { useUpdateUser } from './useUpdateUser'
+export { useUpdateUsersStatus } from './useUpdateUsersStatus'